+++ server.js
@@ -1,7 +1,7 @@
 const express = require("express");
 
 // const mongoose = require("mongoose");
-const nodemailer = require("nodemailer");
+
 // const routes = require("./routes");
 const app = express();
 const PORT = process.env.PORT || 3001;
@@ -21,30 +21,7 @@ if (process.env.NODE_ENV === "production") {
 // mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/kdevproject");
 
 
-//////////////////////NODEMAILER//////////////////
-var transporter = nodemailer.createTransport({
-  service: 'gmail',
-  auth: {
-    user: 'kdev.spring2018@gmail.com',
-    pass: 'gtatl20180508fsf'
-  }
-});
 
-var mailOptions = {
-  from: 'kdev.spring2018@gmail.com',
-  to: 'laeirene59@gmail.com',
-  subject: 'Sending Email using Node.js',
-  text: 'That was easy!'
-};
-
-transporter.sendMail(mailOptions, function(error, info){
-  if (error) {
-    console.log(error);
-  } else {
-    console.log('Email sent: ' + info.response);
-  }
-});
-//////////////////////NODEMAILER//////////////////
